diy/requirements.txt
pip>=9.0
requests
tornado
pyotp
sqlalchemy
suds-py3k
pycrypto